requirements.txt
requests~=2.26.0
python-dotenv~=0.18.0
pymongo~=3.12.0